Set nav graph programatically
