+++ app/src/main/java/com/paulrybitskyi/gamedge/dashboard/DashboardAdapterFragmentFactoryImpl.kt
@@ -1,6 +1,6 @@
 package com.paulrybitskyi.gamedge.dashboard
 
-import com.paulrybitskyi.gamedge.feature.dashboard.adapter.DashboardAdapterFragmentFactory
+import com.paulrybitskyi.gamedge.feature.dashboard.fragment.adapter.DashboardAdapterFragmentFactory
 import com.paulrybitskyi.gamedge.feature.discovery.GamesDiscoveryFragment
 import com.paulrybitskyi.gamedge.feature.likes.LikedGamesFragment
 import com.paulrybitskyi.gamedge.feature.news.GamingNewsFragment
@@ -8,7 +8,8 @@ import com.paulrybitskyi.hiltbinder.BindType
 import javax.inject.Inject
 
 @BindType(installIn = BindType.Component.FRAGMENT)
-internal class DashboardAdapterFragmentFactoryImpl @Inject constructor() : DashboardAdapterFragmentFactory {
+internal class DashboardAdapterFragmentFactoryImpl @Inject constructor() :
+    DashboardAdapterFragmentFactory {
 
     override fun createNewsFragment() = GamingNewsFragment()
     override fun createDiscoveryFragment() = GamesDiscoveryFragment()
