+++ feature-dashboard/build.gradle.kts
@@ -35,6 +35,7 @@ dependencies {
 
     implementation(deps.kotlin.coroutines)
 
+    implementation(deps.androidX.navFragmentKtx)
     implementation(deps.androidX.constraintLayout)
     implementation(deps.androidX.viewPager2)
     implementation(deps.androidX.fragmentKtx)
