+++ feature-dashboard/src/main/res/layout/activity_dashboard.xml
@@ -23,5 +23,4 @@
     android:name="androidx.navigation.fragment.NavHostFragment"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    app:defaultNavHost="true"
-    app:navGraph="@navigation/dashboard_graph"/>
+    app:defaultNavHost="true"/>
